Remove message_parser, move all functions in utils
